Move demoscene.views.generic to common.views
