+++ demoscene/views/groups.py
@@ -9,12 +9,11 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 
-from common.views import writeable_site_required
+from common.views import AjaxConfirmationView, writeable_site_required
 from demoscene.forms.releaser import CreateGroupForm, GroupMembershipForm, GroupSubgroupForm
 from demoscene.models import Edit, Membership, Nick, Releaser
 from demoscene.shortcuts import get_page
 from common.utils.pagination import PaginationControls
-from demoscene.views.generic import AjaxConfirmationView
 
 
 def index(request):
