+++ parties/views/parties.py
@@ -16,11 +16,10 @@ from django.utils.html import format_html
 
 from comments.forms import CommentForm
 from comments.models import Comment
-from common.views import writeable_site_required
+from common.views import AjaxConfirmationView, writeable_site_required
 from demoscene.models import Edit
 from demoscene.shortcuts import simple_ajax_form
 from common.utils.ajax import request_is_ajax
-from demoscene.views.generic import AjaxConfirmationView
 from parties.forms import (
     CompetitionForm, EditPartyForm, EditPartySeriesForm, PartyEditNotesForm, PartyExternalLinkFormSet, PartyForm,
     PartyInvitationFormset, PartyOrganiserForm, PartyReleaseFormset, PartySeriesExternalLinkFormSet,
